+++ src/texture.h
@@ -70,13 +70,13 @@ public:
     static std::shared_ptr<Texture2d> Create(const std::string &key, const std::string &filepath);
     const Vector3 get(const Vector2 &uv) const override;
 
-    void setGridSizeU(unsigned int value);
-    void setGridSizeV(unsigned int value);
+    void setGridSizeU(float value);
+    void setGridSizeV(float value);
 
 protected:
     Texture2d(Image &image);
     Image image;
-    unsigned int gridSizeU;
-    unsigned int gridSizeV;
+    float gridSizeU;
+    float gridSizeV;
 };
